requirements.txt
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.1
async-timeout==4.0.3
certifi==2026.2.25
charset-normalizer==3.4.5
Django==4.2.29
djangorestframework==3.16.1
exceptiongroup==1.3.1
faiss-cpu==1.13.0
filelock==3.19.1
fsspec==2025.10.0
h11==0.16.0
hf-xet==1.4.1
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==0.36.2
idna==3.11
Jinja2==3.1.6
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.28
langchain-core==0.3.83
langchain-text-splitters==0.3.11
langsmith==0.4.37
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.2.1
numpy==2.0.2
orjson==3.11.5
packaging==25.0
pillow==11.3.0
pydantic==2.12.5
pydantic_core==2.41.5
pypdf==6.8.0
python-dotenv==1.2.1
PyYAML==6.0.3
regex==2026.1.15
requests==2.32.5
requests-toolbelt==1.0.0
safetensors==0.7.0
scikit-learn==1.6.1
scipy==1.13.1
sentence-transformers==5.1.2
SQLAlchemy==2.0.48
sqlparse==0.5.5
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tokenizers==0.22.2
torch==2.8.0
tqdm==4.67.3
transformers==4.57.6
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.14.1
zstandard==0.25.0
